refactor; reorg data classes
