+++ cognite_toolkit/_cdf_tk/data_classes/_built_resources.py
@@ -5,7 +5,7 @@ from collections.abc import Collection, Iterator, MutableSequence
 from dataclasses import dataclass
 from functools import cached_property
 from pathlib import Path
-from typing import Any, Generic, SupportsIndex, cast, overload
+from typing import TYPE_CHECKING, Any, Generic, SupportsIndex, cast, overload
 
 from cognite.client.data_classes._base import (
     T_CogniteResourceList,
@@ -14,7 +14,7 @@ from cognite.client.data_classes._base import (
 )
 
 from cognite_toolkit._cdf_tk.exceptions import ToolkitMissingResourceError
-from cognite_toolkit._cdf_tk.loaders import ResourceTypes, get_loader
+from cognite_toolkit._cdf_tk.loaders import get_loader
 from cognite_toolkit._cdf_tk.loaders._base_loaders import T_ID, ResourceLoader, T_WritableCogniteResourceList
 from cognite_toolkit._cdf_tk.utils import (
     calculate_directory_hash,
@@ -24,6 +24,9 @@ from cognite_toolkit._cdf_tk.utils import (
 )
 
 from ._build_variables import BuildVariables
+
+if TYPE_CHECKING:
+    from ._built_modules import BuiltModule
 
 
 @dataclass
@@ -211,70 +214,3 @@ class BuiltFullResourceList(BuiltResourceList[T_ID]):
         if isinstance(index, slice):
             return BuiltFullResourceList[T_ID](super().__getitem__(index))
         return cast(BuiltResourceFull[T_ID], super().__getitem__(index))
-
-
-@dataclass
-class BuiltModule:
-    name: str
-    location: SourceLocation
-    build_variables: BuildVariables
-    resources: dict[str, BuiltResourceList]
-    warning_count: int
-    status: str
-
-    @classmethod
-    def load(cls, data: dict[str, Any]) -> BuiltModule:
-        return cls(
-            name=data["name"],
-            location=SourceLocation.load(data["location"]),
-            build_variables=BuildVariables.load(data["build_variables"]),
-            resources={
-                key: BuiltResourceList([BuiltResource.load(resource_data, key) for resource_data in resources_data])
-                for key, resources_data in data["resources"].items()
-            },
-            warning_count=data.get("warning_count", 0),
-            status=data.get("status", "Success"),
-        )
-
-    def dump(self) -> dict[str, Any]:
-        return {
-            "name": self.name,
-            "location": self.location.dump(),
-            "build_variables": self.build_variables.dump(),
-            "resources": {
-                key: [resource.dump(key) for resource in resources] for key, resources in self.resources.items()
-            },
-            "warning_count": self.warning_count,
-            "status": self.status,
-        }
-
-
-@dataclass
-class BuiltModuleList(list, MutableSequence[BuiltModule]):
-    # Implemented to get correct type hints
-    def __init__(self, collection: Collection[BuiltModule] | None = None) -> None:
-        super().__init__(collection or [])
-
-    def __iter__(self) -> Iterator[BuiltModule]:
-        return super().__iter__()
-
-    @overload
-    def __getitem__(self, index: SupportsIndex) -> BuiltModule: ...
-
-    @overload
-    def __getitem__(self, index: slice) -> BuiltModuleList: ...
-
-    def __getitem__(self, index: SupportsIndex | slice, /) -> BuiltModule | BuiltModuleList:
-        if isinstance(index, slice):
-            return BuiltModuleList(super().__getitem__(index))
-        return super().__getitem__(index)
-
-    def get_resources(self, id_type: type[T_ID], resource_dir: ResourceTypes, kind: str) -> BuiltFullResourceList[T_ID]:
-        return BuiltFullResourceList[T_ID](
-            [
-                resource.create_full(module, resource_dir)
-                for module in self
-                for resource in module.resources.get(resource_dir, [])
-                if resource.kind == kind
-            ]
-        )
